+++ testutil/Fixture/template/declared-layout-view.html.php
@@ -1,0 +1,9 @@
+<?php
+
+use SubstancePHP\HTTP\Renderer\HtmlRenderer;
+
+/** @var HtmlRenderer $this */
+?>
+
+<?php $this->layout('site', ['title' => 'My Site']); ?>
+<p>view body</p>